# Improved allocations and throughput for broadcast scenarios