- Don't allocate when enumerating connections
- Don't allocate tasks unless we truly go async
- Don't get the timestamp, just write the pings always (if there's no ongoing write)
- Track the time since last keep alive write instead of the last write (this needs some reviewing)

## Old

![image](https://user-images.githubusercontent.com/95136/37571804-d1745c1c-2abe-11e8-8e3a-904bc6b6d8bb.png)

```
       Method | Connections | Protocol |         Mean |       Error |        StdDev |      Op/s |   Gen 0 |  Gen 1 | Allocated |
------------- |------------ |--------- |-------------:|------------:|--------------:|----------:|--------:|-------:|----------:|
 SendAsyncAll |           1 |     json |     7.545 us |   0.4048 us |     1.1936 us | 132,532.1 |  0.3204 |      - |   6.62 KB |
 SendAsyncAll |           1 |  msgpack |     5.492 us |   0.2845 us |     0.8343 us | 182,092.0 |  0.0610 |      - |   1.53 KB |
 SendAsyncAll |          10 |     json |    24.614 us |   1.3042 us |     3.7421 us |  40,626.6 |  0.2747 |      - |   7.27 KB |
 SendAsyncAll |          10 |  msgpack |    17.706 us |   0.3166 us |     0.3769 us |  56,476.9 |  0.0916 |      - |   2.17 KB |
 SendAsyncAll |        1000 |     json | 3,620.429 us | 371.4354 us | 1,095.1858 us |     276.2 | 11.7188 | 3.9063 |  593.6 KB |
 SendAsyncAll |        1000 |  msgpack | 2,705.518 us | 103.2645 us |   291.2594 us |     369.6 | 15.6250 | 3.9063 | 516.87 KB |
```

## New

![image](https://user-images.githubusercontent.com/95136/37569110-e9963654-2a9a-11e8-94ca-0557912d8447.png)

```
        Method | Connections | Protocol |         Mean |       Error |      StdDev |       Median |      Op/s |   Gen 0 |  Gen 1 | Allocated |
------------- |------------ |--------- |-------------:|------------:|------------:|-------------:|----------:|--------:|-------:|----------:|
 SendAsyncAll |           1 |     json |     6.614 us |   0.5787 us |   1.7064 us |     6.025 us | 151,184.7 |  0.3052 |      - |   6.33 KB |
 SendAsyncAll |           1 |  msgpack |     3.705 us |   0.1580 us |   0.4585 us |     3.576 us | 269,920.6 |  0.0496 |      - |   1.24 KB |
 SendAsyncAll |          10 |     json |    16.866 us |   0.9062 us |   2.6292 us |    17.083 us |  59,291.6 |  0.2899 | 0.0153 |   6.73 KB |
 SendAsyncAll |          10 |  msgpack |    14.836 us |   0.5637 us |   1.6443 us |    14.572 us |  67,403.7 |  0.0458 |      - |   1.36 KB |
 SendAsyncAll |        1000 |     json | 2,291.156 us |  75.0624 us | 212.9395 us | 2,282.967 us |     436.5 | 11.7188 | 3.9063 | 404.08 KB |
 SendAsyncAll |        1000 |  msgpack | 2,310.744 us | 116.0950 us | 333.0982 us | 2,209.508 us |     432.8 |  7.8125 |      - | 262.68 KB |
```