@davidfowl 2018-03-20
I think we should go through and ValueTask-ify these APIs and re-measure.