@BrennanConroy 2018-03-21
MessagePack allocations went up slightly in the latest, how?