@davidfowl 2018-03-21
> MessagePack allocations went up slightly in the latest, how?

I doubt it's consistent. I've seen strange things with these broadcast tests seems like it's possible to allocate more because more work is happening. If you look at the two memory profiles the `byte[]` allocated in each of the cases (exactly 1000 messages) are slightly different.